Saving to local app storage
Friends can now be saved to local app storage. For this reason, there is a new option to load the default friends from the storage and replace your current friends list with them in the debug section. Deleting everything works but makes the app crash currently.

This is mainly to experiment with modifying friends, which should persist between app restarts, especially notes and quotes.